Fix comment about Get creating a new cookie.
Regarding the comment: "If the named cookie does not exist, this method creates a new cookie with that name."  That is misleading because it only happens if you are using the Get method on the Cookies collection of the Response object.  In the Request object, Get returns null.
I wasted some time on this and based on the posts around the net, I bet dozens more have as well.